+++ tests/test_plugins_os_windows_locale.py
@@ -1,11 +1,5 @@
-import textwrap
-from io import BytesIO
-
 from dissect.target.helpers.regutil import VirtualKey, VirtualValue
-from dissect.target.plugins.os.unix.locale import LocalePlugin as UnixLocalePlugin
 from dissect.target.plugins.os.windows.locale import LocalePlugin as WindowsLocalePlugin
-
-from ._utils import absolute_path
 
 
 def test_locale_plugin_windows(target_win_users, hive_hku, hive_hklm):
@@ -49,22 +43,3 @@ def test_locale_plugin_windows(target_win_users, hive_hku, hive_hklm):
     assert keyboard[0].id == "00000413"
     assert keyboard[1].layout == "us"
     assert keyboard[1].id == "00000409"
-
-
-def test_locale_plugin_unix(target_unix_users, fs_unix):
-    # Locale locations originate from Ubuntu 20.
-
-    fs_unix.map_file_fh("/etc/timezone", BytesIO(textwrap.dedent("Europe/Amsterdam").encode()))
-    fs_unix.map_file_fh("/etc/default/locale", BytesIO(textwrap.dedent("LANG=en_US.UTF-8").encode()))
-    fs_unix.map_file("/etc/default/keyboard", absolute_path("data/unix/configs/keyboard"))
-    target_unix_users.add_plugin(UnixLocalePlugin)
-
-    assert target_unix_users.timezone == "Europe/Amsterdam"
-    assert target_unix_users.language == ["en-US"]
-    keyboard = list(target_unix_users.keyboard())
-    assert len(keyboard) == 1
-    assert keyboard[0].layout == "us"
-    assert keyboard[0].model == "pc105"
-    assert keyboard[0].variant == ""
-    assert keyboard[0].options == ""
-    assert keyboard[0].backspace == "guess"
